# [Inductor] Further tune block size for templated attention on H100